Employee Create

+++ EPMS.Web/Areas/HR/Controllers/DepartmentController.cs
@@ -46,7 +46,7 @@ namespace EPMS.Web.Areas.HR.Controllers
             });
         }
 
-       public ActionResult AddEdit(int? id)
+       public ActionResult AddEdit(long? id)
         {
             DepartmentViewModel viewModel = new DepartmentViewModel();
             if (id != null)
